提交

+++ main_1ppt_ppt_zip.py
@@ -45,13 +45,13 @@ def replace_text_in_slide(slide, old_text, new_text):
 
 if __name__ == '__main__':
     # 第一步：解压压缩包
-    # zip_root_dir = "E:\\workspace\\www.1ppt.com\\www.zip_1ppt.com"
+    # zip_root_dir = "F:\\workspace\\www.1ppt.com\\www.zip_1ppt.com"
     # zip_dirs = sorted(os.listdir(zip_root_dir))
     # zip_files = sorted(os.listdir(zip_root_dir))
     # for zip_file in zip_files:
     #
     #     zip_file_path = zip_root_dir + "\\" + zip_file
-    #     dst_file_path = "E:\\workspace\\www.1ppt.com\\www.uncompress_zip_1ppt.com"
+    #     dst_file_path = "F:\\workspace\\www.1ppt.com\\www.uncompress_zip_1ppt.com"
     #     print("==========" + "开始解压" + zip_file_path + "==========")
     #     try:
     #         # 删除文件名中含有“图”字样文件
@@ -90,7 +90,7 @@ if __name__ == '__main__':
     # exit()
 
     # 第二步：将文件夹中文件移出，并更改文件名称
-    # root_dir = "E:\\workspace\\www.1ppt.com\\www.uncompress_zip_1ppt.com"
+    # root_dir = "F:\\workspace\\www.1ppt.com\\www.uncompress_zip_1ppt.com"
     # files = sorted(os.listdir(root_dir))
     # for file in files:
     #     file_path = root_dir + "\\" + file
@@ -113,8 +113,8 @@ if __name__ == '__main__':
     # exit()
 
     # 第三步：替换幻灯片中文字
-    root_dir = "E:\\workspace\\www.1ppt.com\\www.uncompress_zip_1ppt.com"
-    finish_dir = "E:\\workspace\\www.1ppt.com\\www.finish_zip_1ppt.com"
+    root_dir = "F:\\workspace\\www.1ppt.com\\www.uncompress_zip_1ppt.com"
+    finish_dir = "F:\\workspace\\www.1ppt.com\\www.finish_zip_1ppt.com"
     if not os.path.exists(finish_dir):
         os.makedirs(finish_dir)
     files = sorted(os.listdir(root_dir))
@@ -122,23 +122,25 @@ if __name__ == '__main__':
         file_path = root_dir + "\\" + file
         print(file_path)
 
-        # 获取pptx文件总页数
-        pptx_pages_count = get_slide_count(file_path)
-        if pptx_pages_count <= 5:
-            print(f"幻灯片数量（页数）: {pptx_pages_count}少于5，删除文件")
-            os.remove(file_path)
-            continue
-
-        # 加载PPTX文件
-        prs = Presentation(file_path)
-        # 遍历所有幻灯片并替换文本
-        for slide in prs.slides:
-            # 替换文本
-            replace_text_in_slide(slide, '第一PPT', 'XXXX')
-        # 删除最后一页
-        rId = prs.slides._sldIdLst[-1].rId
-        prs.part.drop_rel(rId)
-        del prs.slides._sldIdLst[-1]
-        # 保存修改后的PPTX文件
         finish_file_path = file_path.replace("uncompress", "finish")
-        prs.save(finish_file_path)
+        if not os.path.exists(finish_file_path):
+            # 获取pptx文件总页数
+            pptx_pages_count = get_slide_count(file_path)
+            if pptx_pages_count <= 5:
+                print(f"幻灯片数量（页数）: {pptx_pages_count}少于5，删除文件")
+                os.remove(file_path)
+                continue
+
+            # 加载PPTX文件
+            prs = Presentation(file_path)
+            # 遍历所有幻灯片并替换文本
+            for slide in prs.slides:
+                # 替换文本
+                replace_text_in_slide(slide, '第一PPT', 'XXXX')
+            # 删除最后一页
+            rId = prs.slides._sldIdLst[-1].rId
+            prs.part.drop_rel(rId)
+            del prs.slides._sldIdLst[-1]
+            # 保存修改后的PPTX文件
+
+            prs.save(finish_file_path)
